Tips To Prepare For A Gaming Job: a blog around tips and tricks in gaming industry
Are you looking for a job in gaming? I was too, and found that the process was difficult. I had several phone and email interviews before landing a job. I'm probably not the only person to share this experience, which is why we started GigsNGaming.com – India's first job board exclusively for gaming professionals.

You can easily find gaming industry jobs at Playjobs. 

With an eminent rise in the Gaming industry and substantial growth in India's gaming market, an exponentially growing number of gamers has appeared on the horizon. Partly due to the same reason is an increasing job market for gaming professionals. Due to this unprecedented situation, opportunities ranging from junior positions as game tester to that of Game director are now present in India. 

An exciting career path awaits you if you are one of those into games, gaming and designing characters. However, now comes the question of preparation for the job you got offered. How do you prepare yourself for a game job? What should you do to get on top of your starting position?

How best to show off your expertise, while also highlighting your ability to learn new games quickly? Believe it or not, we've sorted out just the thing for you. You see, our team has accumulated experience conducting interviews in this field. 

It’s time to get more serious about your preparations if you're planning on joining the gaming industry. As you know, every position in this industry would require people with unique sets of skills that are difficult to find. If you try to apply without possessing them, you will be out of the race long before it begins. You absolutely need a one-of-a-kind portfolio to impress hiring managers, and you can't really expect this much from yourself unless you practice a lot.